+++ src/main.cpp
@@ -1,33 +1,32 @@
 #include <Geode/Geode.hpp>
 #include <Geode/modify/EditorUI.hpp>
 #include <Geode/modify/EditorPauseLayer.hpp>
+#include "main.hpp"
 
 using namespace geode::prelude;
 
-struct TriggerData {
-    EffectGameObject* trigger;
-    int targetGroup = 0;
-    int centerGroup = 0;
-
-    TriggerData(EffectGameObject* obj, int target, int center)
-        : trigger(obj), targetGroup(target), centerGroup(center) {}
-};
-
-std::unordered_map<int, int> groupParentObjs;
-
-auto mod = Mod::get();
-
-bool copyParent = mod->getSavedValue<bool>("copy-parent-enabled");
-
 class $modify(Editor, EditorUI) {
+    bool init(LevelEditorLayer* editorLayer) {
+        if (!EditorUI::init(editorLayer)) return false;
+        auto mod = Mod::get();
+        copyParent = mod->getSavedValue<bool>("copy-parent-enabled");
+        replaceBuildHelper = mod->getSavedValue<bool>("replace-build-helper");
+        return true;
+    }
+
+    void dynamicGroupUpdate(bool p0) {
+        if (replaceBuildHelper) this->buildHelperExtra(this->getSelectedObjects());
+        else EditorUI::dynamicGroupUpdate(p0);
+    }
+
     gd::string copyObjects(CCArray* objects, bool copyColors, bool sort) {
         // ok so this entire thing is built on the assumption that the copy gdstring builds based of the order of the inputted ccarray
         // i have no idea why this wouldnt be the case and in testing it works like that but this is also rly sketchy, idk if it works it works
         groupParentObjs.clear();
         auto parentDict = m_editorLayer->m_parentGroupsDict;
         for (auto key : CCArrayExt<CCInteger*>(parentDict->allKeys())) {
-            auto value = key->getValue(); // allkeys returns ccint for some ungodly reason even tho objectforkey takes int not ccint why cocos
-            auto index = objects->indexOfObject(parentDict->objectForKey(value)); // fuck null checking robtop got me
+            int value = key->getValue(); // allkeys returns ccint for some ungodly reason even tho objectforkey takes int not ccint why cocos
+            int index = objects->indexOfObject(parentDict->objectForKey(value)); // fuck null checking robtop got me
             if (index != UINT_MAX) groupParentObjs.insert({value, index}); // fuck null checking indexofobject returning uint_max on null got me
         }
         return EditorUI::copyObjects(objects, copyColors, sort);
@@ -49,7 +48,7 @@ class $modify(Editor, EditorUI) {
     void pasteBuildHelper() {
         copyParent = false;
         auto objs = pasteObjects(copyObjects(getSelectedObjects(), false, false), false, false);
-        copyParent = mod->getSavedValue<bool>("copy-parent-enabled");
+        copyParent = Mod::get()->getSavedValue<bool>("copy-parent-enabled");
         buildHelperExtra(objs, true);
         this->updateButtons();
         this->updateObjectInfoLabel();
@@ -61,8 +60,8 @@ class $modify(Editor, EditorUI) {
         auto objs = ccArrayToVector<GameObject*>(array);
         for (auto obj : objs) {
             if (auto trigger = typeinfo_cast<EffectGameObject*>(obj)) {
-                auto target = trigger->m_targetGroupID;
-                auto center = trigger->m_centerGroupID;
+                int target = trigger->m_targetGroupID;
+                int center = trigger->m_centerGroupID;
                 
                 triggerGroups.insert(target);
                 triggerGroups.insert(center);
@@ -131,7 +130,7 @@ class $modify(EditorPause, EditorPauseLayer) {
     }
 
     void onToggleCopyParent(CCObject* sender) {
-        mod->setSavedValue<bool>("copy-parent-enabled", !copyParent);
+        Mod::get()->setSavedValue<bool>("copy-parent-enabled", !copyParent);
         copyParent = !copyParent;;
     }