+++ changelog.md
@@ -1,2 +1,7 @@
+## v1.1.0
+- **Added** replace build helper
+
+- **Fixed** technically made the code more "correct" but i also added a main.hpp so still not great
+
 ## v1.0.0
 - Initial Release